Added SQL mdf file
Add the SQL mdf file and refer the MDF connection in the appsettings .json file. This should allow the project read the data from the mdf file and then run the test project to see the rewards.

You can pump in the data into the transaction table by using transaction controller endpoint.